Week7/day1/requirements.txt
openai
faiss-cpu
pandas
python-dotenv
tqdm
PyPDF2
python-docx
pyyaml
tiktoken